Remove copy_hab_stuff
copy_hab_stuff brought in way more than was necessary and allowed the exporter to work when it should have been broken. To resolve this, KERNEL and SYSTEM_PKGS variables were added. These are things needed to boot the system, but aren't necessarily provided by the user packages, and lay the groundwork for custom kernels